Fix timeouts + new Jira Account id requirements (#3)
Bunch of silly stuff in here:

1. **Jira account name problems**

Jira seems to have decided that it will send account IDs instead of jira names from their user API

This means most of the name checking logic is incompatible back and forth between slack/jira, and the account name regex did not work 

So: 

The regex has been updated to support `:` and `-`

It also then makes sure to swap the `"[~accountid:#####]"` user name format with the jira short name format so that way the comments from the bot contain the actual understandable names.

2. **Timeouts and duplicates**

For some reason the foreach index was undefined so we never actually returned the call for heroku, causing it to continually try. 

I didn't try to figure out why that happens I just made my own index instead. 

**CAVEAT & CURRENT MISSING FEATURE**

I had to manually manually go into mongo and change the format of the usernames. Here's an example: 

BEFORE:

```
{
    "_id": {
        "$oid": "#######"
    },
    "slackUsername": "johnny",
    "__v": 0,
    "jiraToken": "FOO",
    "jiraTokenSecret": "BAR",
    "jiraUsername":  "[~john]"
}
```

AFTER

```
{
    "_id": {
        "$oid": "#######"
    },
    "slackUsername": "johnny",
    "__v": 0,
    "jiraToken": "FOO",
    "jiraTokenSecret": "BAR",
    "jiraUsername": "[~accountid:######]",
    "jiraShortName": "[~john]"
}
```

So then we use the userame when jira talks to us, and use the short name to talk to slack and jira